python-services/document-processing-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
markdown==3.5.1
langchain==0.0.340
langchain-community==0.0.1

# Vector Database
chromadb==0.4.18

# Message Queue
pika==1.3.2
celery==5.3.4

# Service Discovery
nacos-sdk-python==0.1.9

# Monitoring
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-exporter-zipkin==1.21.0

# Utilities
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.25.2
redis==5.0.1
